Add files via upload
A cafe billing system built on a GUI. The menu card that shows up as the interface takes orders, bills out the total amount to the user, and asks the user's preference for payment. It is made using the concepts of object-oriented programming. Java is the language of choice because it works with all hardware and is excellent for comprehending OOPS concepts.